requirements.txt
beautifulsoup4
newspaper3k
